+++ case-studies/higgs/dataframe/analyzers/CMakeLists.txt
@@ -13,17 +13,26 @@ get_target_property(EDM4HEP_INCLUDE_DIRS EDM4HEP::edm4hep INTERFACE_INCLUDE_DIRE
 
 include_directories(${EDM4HEP_INCLUDE_DIRS} ${podio_INCLUDE_DIR})
 
-ROOT_GENERATE_DICTIONARY(G__FCCAnalysesHiggs dummyLoaderHiggs.h LINKDEF LinkDef.h)
+ROOT_GENERATE_DICTIONARY(G__FCCAnalysesHiggs 
+                dummyLoaderHiggs.h
+                APCHiggsTools.h
+                LINKDEF LinkDef.h)
 
 
 file(GLOB sources *.cc)
 file(GLOB headers *.h)
 add_library(FCCAnalysesHiggs SHARED ${sources} ${headers} G__FCCAnalysesHiggs.cxx )
 target_include_directories(FCCAnalysesHiggs  PUBLIC
-        ${CMAKE_SOURCE_DIR}/analyzers
-	${FCCANALYSES_INCLUDE_PATH}
+                           ${CMAKE_SOURCE_DIR}/analyzers
+	                         ${FCCANALYSES_INCLUDE_PATH}
 )
-target_link_libraries(FCCAnalysesHiggs ROOT::Physics ROOT::MathCore ROOT::ROOTVecOps ROOT::ROOTDataFrame EDM4HEP::edm4hep EDM4HEP::edm4hepDict podio::podio)
+target_link_libraries(FCCAnalysesHiggs ROOT::Physics 
+                      ROOT::MathCore 
+                      ROOT::ROOTVecOps 
+                      ROOT::ROOTDataFrame 
+                      EDM4HEP::edm4hep 
+                      EDM4HEP::edm4hepDict 
+                      podio::podio)
 
 set_target_properties(FCCAnalysesHiggs PROPERTIES
   PUBLIC_HEADER "${headers}")
