--- case-studies/higgs/dataframe/analyzers/ReconstructedParticle_caseStudy_mHRecoil.h
@@ -1,139 +1,0 @@
-
-#ifndef  RECONSTRUCTEDPARTICLE_ANALYZERS_H
-#define  RECONSTRUCTEDPARTICLE_ANALYZERS_H
-
-#include <cmath>
-#include <vector>
-
-#include "TLorentzVector.h"
-#include "ROOT/RVec.hxx"
-#include "edm4hep/ReconstructedParticleData.h"
-#include "edm4hep/ParticleIDData.h"
-
-namespace ReconstructedParticle{
-	///build the resonance from 2 particles from an arbitrary list of input ReconstructedPartilces. Keep the closest to the mass given as input
-	struct resonanceZBuilder {
-		float m_resonance_mass;
-		resonanceZBuilder(float arg_resonance_mass);
-		ROOT::VecOps::RVec<edm4hep::ReconstructedParticleData> operator()(ROOT::VecOps::RVec<edm4hep::ReconstructedParticleData> legs);
-	};
-
-  /// build the resonance from 2 particles from an arbitrary list of input ReconstructedPartilces. Keep the closest to the mass given as input
-  struct resonanceBuilder {
-    float m_resonance_mass;
-    resonanceBuilder(float arg_resonance_mass);
-    ROOT::VecOps::RVec<edm4hep::ReconstructedParticleData> operator()(ROOT::VecOps::RVec<edm4hep::ReconstructedParticleData> legs);
-  };
-  
-  /// build the recoil from an arbitrary list of input ReconstructedPartilces and the center of mass energy
-  struct recoilBuilder {
-    recoilBuilder(float arg_sqrts);
-    float m_sqrts = 240.0;
-    ROOT::VecOps::RVec<edm4hep::ReconstructedParticleData> operator() (ROOT::VecOps::RVec<edm4hep::ReconstructedParticleData> in) ;
-  };
-  
-  /// return the angular separations (min / max / average) between a collection of particles
-  struct angular_separationBuilder {
-    angular_separationBuilder( int arg_delta); //  0, 1, 2 = max, min, average
-    int m_delta = 0;
-    float operator() (ROOT::VecOps::RVec<edm4hep::ReconstructedParticleData> in) ;
-  };
-
-  /// select ReconstructedParticles with transverse momentum greater than a minimum value [GeV]
-  struct sel_pt {
-    sel_pt(float arg_min_pt);
-    float m_min_pt = 1.; //> transverse momentum threshold [GeV]
-    ROOT::VecOps::RVec<edm4hep::ReconstructedParticleData>  operator() (ROOT::VecOps::RVec<edm4hep::ReconstructedParticleData> in);
-  };
-  
-  /// select ReconstructedParticles with momentum greater than a minimum value [GeV]
-  struct sel_p {
-    sel_p(float arg_min_p);
-    float m_min_p = 1.; //> momentum threshold [GeV]
-    ROOT::VecOps::RVec<edm4hep::ReconstructedParticleData>  operator() (ROOT::VecOps::RVec<edm4hep::ReconstructedParticleData> in);
-  };
-
-  /// select ReconstructedParticles with charge equal or in asolute value
-  struct sel_charge {
-    sel_charge(int arg_charge, bool arg_abs);
-    float m_charge; //> charge condition
-    bool  m_abs;//> absolute value of the charge
-    ROOT::VecOps::RVec<edm4hep::ReconstructedParticleData>  operator() (ROOT::VecOps::RVec<edm4hep::ReconstructedParticleData> in);
-  };
-
-  /// select a list of reconstructed particles depending on the angle cosTheta axis
-  struct sel_axis{
-    bool m_pos = 0; //> Which hemisphere to select, false/0=cosTheta<0 true/1=cosTheta>0
-    sel_axis(bool arg_pos);
-    ROOT::VecOps::RVec<edm4hep::ReconstructedParticleData> operator()(ROOT::VecOps::RVec<float> angle, ROOT::VecOps::RVec<edm4hep::ReconstructedParticleData> in);
-  };
-
-  
-  
-  
-  /// return reconstructed particles
-  ROOT::VecOps::RVec<edm4hep::ReconstructedParticleData> get(ROOT::VecOps::RVec<int> index, ROOT::VecOps::RVec<edm4hep::ReconstructedParticleData> in);
-
-  /// return the transverse momenta of the input ReconstructedParticles
-  ROOT::VecOps::RVec<float> get_pt(ROOT::VecOps::RVec<edm4hep::ReconstructedParticleData> in);
-  
-  /// return the momenta of the input ReconstructedParticles
-  ROOT::VecOps::RVec<float> get_p(ROOT::VecOps::RVec<edm4hep::ReconstructedParticleData> in);
-  
-  /// return the momenta of the input ReconstructedParticles
-  ROOT::VecOps::RVec<float> get_px(ROOT::VecOps::RVec<edm4hep::ReconstructedParticleData> in);
-
-  /// return the momenta of the input ReconstructedParticles
-  ROOT::VecOps::RVec<float> get_py(ROOT::VecOps::RVec<edm4hep::ReconstructedParticleData> in);
-
-  /// return the momenta of the input ReconstructedParticles
-  ROOT::VecOps::RVec<float> get_pz(ROOT::VecOps::RVec<edm4hep::ReconstructedParticleData> in);
-  
-  /// return the pseudo-rapidity of the input ReconstructedParticles
-  ROOT::VecOps::RVec<float> get_eta(ROOT::VecOps::RVec<edm4hep::ReconstructedParticleData> in);
-
-  /// return the rapidity of the input ReconstructedParticles
-  ROOT::VecOps::RVec<float> get_y(ROOT::VecOps::RVec<edm4hep::ReconstructedParticleData> in);
-
-  /// return the theta of the input ReconstructedParticles
-  ROOT::VecOps::RVec<float> get_theta(ROOT::VecOps::RVec<edm4hep::ReconstructedParticleData> in);
-
-  /// return the phi of the input ReconstructedParticles
-  ROOT::VecOps::RVec<float> get_phi(ROOT::VecOps::RVec<edm4hep::ReconstructedParticleData> in);
-
-  /// return the energy of the input ReconstructedParticles
-  ROOT::VecOps::RVec<float> get_e(ROOT::VecOps::RVec<edm4hep::ReconstructedParticleData> in);
-  
-  /// return the masses of the input ReconstructedParticles
-  ROOT::VecOps::RVec<float> get_mass(ROOT::VecOps::RVec<edm4hep::ReconstructedParticleData> in); 
-
-  /// return the charges of the input ReconstructedParticles
-  ROOT::VecOps::RVec<float> get_charge(ROOT::VecOps::RVec<edm4hep::ReconstructedParticleData> in); 
-
-  /// return the TlorentzVector of the input ReconstructedParticles
-  ROOT::VecOps::RVec<TLorentzVector> get_tlv(ROOT::VecOps::RVec<edm4hep::ReconstructedParticleData> in);
-
-  /// concatenate both input vectors and return the resulting vector
-  ROOT::VecOps::RVec<edm4hep::ReconstructedParticleData> merge(ROOT::VecOps::RVec<edm4hep::ReconstructedParticleData> x, ROOT::VecOps::RVec<edm4hep::ReconstructedParticleData> y);
-  
-  /// return the size of the input collection
-  int get_n(ROOT::VecOps::RVec<edm4hep::ReconstructedParticleData> in);
-
-  /// returns the bjet flavour
-  ROOT::VecOps::RVec<bool> getJet_btag(ROOT::VecOps::RVec<int> index, ROOT::VecOps::RVec<edm4hep::ParticleIDData> pid, ROOT::VecOps::RVec<float> values); 
-
-  /// get number of b-jets
-  int getJet_ntags(ROOT::VecOps::RVec<bool> in);
-
-
-	/// return the costheta of the input ReconstructedParticles
-	ROOT::VecOps::RVec<float> get_costheta(ROOT::VecOps::RVec<edm4hep::ReconstructedParticleData> in);
-
-	/// return the costheta of the input ReconstructedParticles MET
-  //ROOT::VecOps::RVec<float> get_MET_costheta(ROOT::VecOps::RVec<Float_t> Px, ROOT::VecOps::RVec<Float_t> Py, ROOT::VecOps::RVec<Float_t> Pz);
-  //float get_MET_costheta(Float_t Px, Float_t Py, Float_t Pz);
-  ROOT::VecOps::RVec<float> get_MET_costheta(ROOT::VecOps::RVec<Float_t>Px, ROOT::VecOps::RVec<Float_t>Py, ROOT::VecOps::RVec<Float_t>Pz, ROOT::VecOps::RVec<Float_t>E);
-  ///return muon_quality_check result
-	ROOT::VecOps::RVec<edm4hep::ReconstructedParticleData>  muon_quality_check(ROOT::VecOps::RVec<edm4hep::ReconstructedParticleData> in);
-}
-#endif
