+++ case-studies/higgs/dataframe/localSetup.sh
@@ -1,2 +1,2 @@
 export LD_LIBRARY_PATH=$PWD/install/lib:$LD_LIBRARY_PATH
-export ROOT_INCLUDE_PATH=$PWD/install/include/FCCAnalysesTop:$ROOT_INCLUDE_PATH
+export ROOT_INCLUDE_PATH=$PWD/install/include/FCCAnalysesHiggs:$ROOT_INCLUDE_PATH
